refactor: [ALT-270] 채팅 리뷰 컨벤션 반영
- GroupChatRoomProvider를 usecase 패키지 밖(chat/support)으로 이동:
  inbound port를 구현하지 않는 내부 협력자이므로 use case 규칙과 분리
- 테스트 클래스명 16개를 파일명(*Tests, 복수)과 일치하도록 *Test → *Tests 통일
