+++ src/main/java/com/dreamteam/alter/application/chat/usecase/SyncWorkspaceChatMembership.java
@@ -5,6 +5,7 @@ import com.dreamteam.alter.domain.chat.entity.ChatRoom;
 import com.dreamteam.alter.domain.chat.entity.ChatRoomMember;
 import com.dreamteam.alter.domain.chat.port.inbound.SyncWorkspaceChatMembershipUseCase;
 import com.dreamteam.alter.domain.chat.port.outbound.*;
+import com.dreamteam.alter.application.chat.support.GroupChatRoomProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
